+++ RELEASES.md
@@ -14,6 +14,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: forms for composite templates (work in progress)
 * prov-template-compiler: composite templates conversion to csv
 * service-templates: form interface to support composite templates (csv or json)
+* service-templates: composite templates are persisted as record linking to composed instances